+++ lightpick.js
@@ -1,5 +1,5 @@
 /**
-* @version: 0.0.1
+* @version: 0.0.2
 * @author: Rinat G. http://coding.kz
 * @copyright: Copyright (c) 2018 Rinat G.
 * @license: Licensed under the MIT license. See http://www.opensource.org/licenses/mit-license.php
@@ -22,6 +22,36 @@
     }
 }(this, function(moment, $) {
     'use strict';
+
+    if (typeof Object.assign != 'function') {
+        // Must be writable: true, enumerable: false, configurable: true
+        Object.defineProperty(Object, "assign", {
+          value: function assign(target, varArgs) { // .length of function is 2
+            'use strict';
+            if (target == null) { // TypeError if undefined or null
+              throw new TypeError('Cannot convert undefined or null to object');
+            }
+      
+            var to = Object(target);
+      
+            for (var index = 1; index < arguments.length; index++) {
+              var nextSource = arguments[index];
+      
+              if (nextSource != null) { // Skip over if undefined or null
+                for (var nextKey in nextSource) {
+                  // Avoid bugs when hasOwnProperty is shadowed
+                  if (Object.prototype.hasOwnProperty.call(nextSource, nextKey)) {
+                    to[nextKey] = nextSource[nextKey];
+                  }
+                }
+              }
+            }
+            return to;
+          },
+          writable: true,
+          configurable: true
+        });
+    }
 
     var document = window.document,
 
@@ -270,7 +300,8 @@
     },
 
     updateDates = function(el, opts){
-        el.querySelectorAll('.day').forEach(function(day) {
+        var days = el.querySelectorAll('.day');
+        [].forEach.call(days, function(day) {
             day.outerHTML = renderDay(opts, parseInt(day.getAttribute('data-time')), false, day.className.split(' '));
         });
     },
@@ -376,8 +407,8 @@
                 if (!hoverDate.isValid()) {
                     return;
                 }
-
-                self.el.querySelectorAll('.day').forEach(function(day) {
+                var days = self.el.querySelectorAll('.day');
+                [].forEach.call(days, function(day) {
                     var dt = moment(parseInt(day.getAttribute('data-time')));
 
                     day.classList.remove('invert');