+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+0.0.2 - June 2018
+* Fix typo in docs
+* Add polyfill (Object.assign)
+* flexbox fallback for IE
+
 0.0.1 - June 2018
-- Published on Github
+* Published on Github